refactor: split AdminPanelV4 into hooks, tabs, and sub-components
Breaks a 1960-line monolith into AdminPanelV4/ directory:
- 6 custom hooks (useAnchors, useLabels, useRoomConfig, useRecording,
  usePlayback, useParticipants) each owning their domain state
- 8 tab components (one per admin tab); ParticipantsTab and MomentsTab
  wrapped in React.memo to avoid re-renders on every cursor move
- ParticipantRow and OfferInterfaceModal extracted as components
- types.ts for shared local types and anchorToLocal() helper
- index.tsx thin orchestrator (~180 lines) with ref-based socket dispatch

No behavior changes; import path is backward-compatible.
